+++ inst/examples/ex-seek.R
@@ -1,0 +1,6 @@
+## Seek columns
+seek_columns(iris, select = startsWith, prefix = "Sepal")
+
+## Get columns
+x <- get_columns(iris, select = startsWith, prefix = "Sepal")
+head(x)